+++ 01. Project/Engine/CMesh.cpp
@@ -34,7 +34,7 @@ int CMesh::Create(void* _pVtxSystem, size_t _iVtxCount, void* _pIdxSysmem, size_
 	HRESULT hr = S_OK;
 
 	// System -> GPU	
-	m_tVBDesc.ByteWidth = sizeof(Vtx) * _iVtxCount;
+	m_tVBDesc.ByteWidth = (UINT)(sizeof(Vtx) * _iVtxCount);
 	m_tVBDesc.BindFlags = D3D11_BIND_FLAG::D3D11_BIND_VERTEX_BUFFER;
 	m_tVBDesc.Usage = D3D11_USAGE_DEFAULT;
 	m_tVBDesc.CPUAccessFlags = 0;
@@ -45,7 +45,7 @@ int CMesh::Create(void* _pVtxSystem, size_t _iVtxCount, void* _pIdxSysmem, size_
 	hr = DEVICE->CreateBuffer(&m_tVBDesc, &tSubData, &m_VB);
 
 	// IndexBuffer ¸¸µé±â	
-	m_iIdxCount = _iIdxCount;
+	m_iIdxCount = (UINT)_iIdxCount;
 	m_tIBDesc.ByteWidth = sizeof(UINT) * m_iIdxCount;
 	m_tIBDesc.BindFlags = D3D11_BIND_FLAG::D3D11_BIND_INDEX_BUFFER;
 	m_tIBDesc.Usage = D3D11_USAGE_DEFAULT;
